requirements/base.txt
Django==4.1
django-environ==0.8.1
djangorestframework==3.13.1
django-cors-headers==3.10.1
django-filter==21.1
django-autoslug==1.9.8
django-countries==7.2.1
django-phonenumber-field==5.2.0
phonenumbers==8.12.33
drf-spectacular==0.23.1

Pillow==9.0.0

argon2-cffi==21.3.0

pytz==2021.3

redis==4.5.1
celery==5.2.7
django-celery-beat==2.4.0
django-celery-email==3.0.0
flower==2.0.1

djoser==2.1.0
djangorestframework-simplejwt==4.8.0
PyJWT==2.1.0